backend/requirements.txt
fastapi
uvicorn
python-dotenv
aiosmtplib
pydantic
torch
torchvision
scikit-learn
Pillow
matplotlib
numpy
tqdm
pandas
joblib
